@@ -8,13 +8,13 @@
   },
   "dependencies": {
     "easy-peasy": "^6.0.0",
-    "next": "14.2.21",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-redux": "7.1.3",
-    "redux": "4.0.5",
-    "redux-devtools-extension": "2.13.8",
-    "redux-thunk": "2.3.0"
+    "next": "15.1.7",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-redux": "9.2.0",
+    "redux": "5.0.1",
+    "redux-devtools-extension": "2.13.9",
+    "redux-thunk": "3.1.0"
   },
   "license": "MIT",
   "keywords": [],
