@@ -57,13 +57,13 @@
     "prepare": "husky install"
   },
   "resolutions": {
-    "@types/react": "^18.0.20"
+    "@types/react": "^19.0.8"
   },
   "peerDependencies": {
-    "@types/react": "^16.8 || ^17.0 || ^18.0",
-    "@types/react-dom": "^16.8 || ^17.0 || ^18.0",
-    "react": "^16.8 || ^17.0 || ^18.0",
-    "react-dom": "^16.8 || ^17.0 || ^18.0",
+    "@types/react": "^18.0 || ^19.0",
+    "@types/react-dom": "^18.0 || ^19.0",
+    "react": "^18.0 || ^19.0",
+    "react-dom": "^18.0 || ^19.0",
     "react-native": ">=0.59"
   },
   "peerDependenciesMeta": {
@@ -84,10 +84,10 @@
     "@babel/runtime": "^7.22.6",
     "fast-deep-equal": "^3.1.3",
     "immer": "^9.0.21",
-    "redux": "^4.1.2",
-    "redux-thunk": "^2.4.1",
+    "redux": "^5.0.1",
+    "redux-thunk": "^3.1.0",
     "ts-toolbelt": "^9.6.0",
-    "use-sync-external-store": "^1.2.0"
+    "use-sync-external-store": "^1.4.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.22.9",
@@ -104,11 +104,12 @@
     "@babel/preset-typescript": "^7.22.5",
     "@babel/register": "^7.22.5",
     "@definitelytyped/dtslint": "^0.0.163",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^14.0.0",
-    "@types/react": "^18.2.15",
-    "@types/react-dom": "^18.2.7",
-    "@types/react-redux": "^7.1.25",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.2.0",
+    "@testing-library/dom": "^10.0.0",
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
+    "@types/react-redux": "^7.1.34",
     "app-root-dir": "^1.0.2",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^29.6.1",
@@ -127,10 +128,10 @@
     "pinst": "^3.0.0",
     "prettier": "^3.0.0",
     "prettier-eslint": "^15.0.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-redux": "^8.0.2",
-    "react-test-renderer": "^18.2.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-redux": "^9.2.0",
+    "react-test-renderer": "^19.0.0",
     "rimraf": "^5.0.1",
     "rollup": "^2.67.1",
     "rollup-plugin-analyzer": "^4.0.0",
@@ -138,7 +139,7 @@
     "rollup-plugin-node-resolve": "^5.2.0",
     "shallowequal": "^1.1.0",
     "title-case": "^3.0.3",
-    "typescript": "5.1.6"
+    "typescript": "5.7.3"
   },
   "jest": {
     "collectCoverageFrom": [
@@ -146,7 +147,7 @@
     ],
     "setupFilesAfterEnv": [
       "<rootDir>/node_modules/regenerator-runtime/runtime",
-      "@testing-library/jest-dom/extend-expect"
+      "@testing-library/jest-dom"
     ],
     "testEnvironment": "jsdom",
     "testMatch": [
