@@ -1,15 +1,16 @@
 import { StoreProvider } from 'easy-peasy';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import App from './App';
 import store from './store';
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root')!);
+
+root.render(
   <React.StrictMode>
     <StoreProvider store={store}>
       <App />
     </StoreProvider>
   </React.StrictMode>,
-  document.getElementById('root'),
 );
